backend/requirements.txt
# Core dependencies
playwright>=1.40.0
redis>=5.0.0
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Settings
pydantic-settings>=2.0.0

# API
fastapi>=0.109.0
uvicorn>=0.25.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation
pydantic[email]>=2.0.0

# Migrations
alembic>=1.13.0

python-telegram-bot
